Refactor code to be more Object Oriented - Move all the minification to an extension to StreamReader - Get rid of global var _features - Extend the Features class with a constructor and add MaxLength
