+++ HtmlMinifier.Tests/HtmlMinifier.Tests.csproj
@@ -75,6 +75,9 @@
     <Content Include="Data\StandardResult.txt" />
     <Content Include="Data\Standard.txt" />
   </ItemGroup>
+  <ItemGroup>
+    <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
