+++ ViewMinifier/HtmlMinifier.csproj
@@ -59,6 +59,7 @@
     <Compile Include="Features.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="StreamReaderExtension.cs" />
     <Compile Include="StringExtension.cs" />
   </ItemGroup>
   <ItemGroup>
